+++ Project/PrivateVault App/Code/UI/Gallery/GalleryView Extensions/GalleryView+NavigationButtons.swift
